pageSetup tests
